# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SMT Feeder Scanning & Verification System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Bom {
  id: number;
  name: string;
  description?: string;
  itemCount: number;
  createdAt: string;
}

export interface BomItem {
  id: number;
  bomId: number;
  feederNumber: string;
  partNumber: string;
  description?: string;
  location?: string;
  quantity: number;
}

export interface BomDetail {
  id: number;
  name: string;
  description?: string;
  items: BomItem[];
  createdAt: string;
}

export interface CreateBomRequest {
  name: string;
  description?: string;
}

export interface UpdateBomRequest {
  name?: string;
  description?: string;
}

export interface CreateBomItemRequest {
  feederNumber: string;
  partNumber: string;
  description?: string;
  location?: string;
  quantity: number;
}

export interface UpdateBomItemRequest {
  feederNumber?: string;
  partNumber?: string;
  description?: string;
  location?: string;
  quantity?: number;
}

export type SessionStatus = (typeof SessionStatus)[keyof typeof SessionStatus];

export const SessionStatus = {
  active: "active",
  completed: "completed",
} as const;

export interface Session {
  id: number;
  bomId: number;
  bomName?: string;
  companyName: string;
  customerName?: string;
  panelName: string;
  supervisorName: string;
  operatorName: string;
  shiftName: string;
  shiftDate: string;
  logoUrl?: string;
  productionCount?: number;
  status: SessionStatus;
  startTime: string;
  endTime?: string;
  createdAt: string;
}

export type SessionDetailStatus =
  (typeof SessionDetailStatus)[keyof typeof SessionDetailStatus];

export const SessionDetailStatus = {
  active: "active",
  completed: "completed",
} as const;

export type ScanRecordStatus =
  (typeof ScanRecordStatus)[keyof typeof ScanRecordStatus];

export const ScanRecordStatus = {
  ok: "ok",
  reject: "reject",
} as const;

export interface ScanRecord {
  id: number;
  sessionId: number;
  feederNumber: string;
  spoolBarcode?: string;
  status: ScanRecordStatus;
  partNumber?: string;
  description?: string;
  location?: string;
  scannedAt: string;
}

export interface SessionDetail {
  id: number;
  bomId: number;
  bomName?: string;
  companyName: string;
  customerName?: string;
  panelName: string;
  supervisorName: string;
  operatorName: string;
  shiftName: string;
  shiftDate: string;
  logoUrl?: string;
  productionCount?: number;
  status: SessionDetailStatus;
  startTime: string;
  endTime?: string;
  scans: ScanRecord[];
  createdAt: string;
}

export interface SpliceRecord {
  id: number;
  sessionId: number;
  feederNumber: string;
  oldSpoolBarcode: string;
  newSpoolBarcode: string;
  splicedAt: string;
  durationSeconds?: number;
}

export interface CreateSpliceRequest {
  feederNumber: string;
  oldSpoolBarcode: string;
  newSpoolBarcode: string;
  durationSeconds?: number;
}

export interface CreateSessionRequest {
  bomId?: number;
  companyName: string;
  customerName?: string;
  panelName: string;
  supervisorName: string;
  operatorName: string;
  shiftName: string;
  shiftDate: string;
  logoUrl?: string;
  productionCount?: number;
}

export type UpdateSessionRequestStatus =
  (typeof UpdateSessionRequestStatus)[keyof typeof UpdateSessionRequestStatus];

export const UpdateSessionRequestStatus = {
  active: "active",
  completed: "completed",
} as const;

export interface UpdateSessionRequest {
  endTime?: string;
  productionCount?: number;
  status?: UpdateSessionRequestStatus;
  logoUrl?: string;
}

export interface ScanFeederRequest {
  feederNumber: string;
  spoolBarcode?: string;
}

export type ScanResultStatus =
  (typeof ScanResultStatus)[keyof typeof ScanResultStatus];

export const ScanResultStatus = {
  ok: "ok",
  reject: "reject",
} as const;

export interface ScanResult {
  scan: ScanRecord;
  status: ScanResultStatus;
  message: string;
}

export interface SessionSummary {
  sessionId: number;
  totalBomItems: number;
  scannedCount: number;
  okCount: number;
  rejectCount: number;
  missingCount: number;
  completionPercent: number;
  durationMinutes?: number;
}

export interface SessionReport {
  session: SessionDetail;
  summary: SessionSummary;
  bomItems: BomItem[];
}

export interface AnalyticsOverview {
  totalSessions: number;
  activeSessions: number;
  completedSessions: number;
  totalScans: number;
  totalOk: number;
  totalReject: number;
  overallOkRate: number;
  totalBoms: number;
  avgDurationMinutes?: number;
}

export interface ParetoItem {
  feederNumber: string;
  partNumber?: string;
  description?: string;
  rejectCount: number;
  cumulativePercent: number;
}

export interface ParetoData {
  items: ParetoItem[];
  totalRejects: number;
}

export interface SessionTrend {
  date: string;
  sessions: number;
  totalScans: number;
  okCount: number;
  rejectCount: number;
  okRate: number;
}

export type GetAnalyticsParetoParams = {
  sessionId?: number;
};