# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SMT Feeder Scanning & Verification System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bom:
    get:
      operationId: listBoms
      tags: [bom]
      summary: List all BOMs
      responses:
        "200":
          description: List of BOMs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bom"
    post:
      operationId: createBom
      tags: [bom]
      summary: Create a new BOM
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBomRequest"
      responses:
        "201":
          description: BOM created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bom"
  /bom/{bomId}:
    get:
      operationId: getBom
      tags: [bom]
      summary: Get a BOM by ID
      parameters:
        - name: bomId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: BOM detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BomDetail"
    patch:
      operationId: updateBom
      tags: [bom]
      summary: Update a BOM
      parameters:
        - name: bomId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBomRequest"
      responses:
        "200":
          description: Updated BOM
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bom"
    delete:
      operationId: deleteBom
      tags: [bom]
      summary: Delete a BOM
      parameters:
        - name: bomId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /bom/{bomId}/items:
    post:
      operationId: addBomItem
      tags: [bom]
      summary: Add a feeder item to BOM
      parameters:
        - name: bomId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBomItemRequest"
      responses:
        "201":
          description: BOM item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BomItem"
  /bom/{bomId}/items/{itemId}:
    patch:
      operationId: updateBomItem
      tags: [bom]
      summary: Update a BOM item
      parameters:
        - name: bomId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBomItemRequest"
      responses:
        "200":
          description: Item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BomItem"
    delete:
      operationId: deleteBomItem
      tags: [bom]
      summary: Delete a BOM item
      parameters:
        - name: bomId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Item deleted
...
components:
  schemas:
...
    Bom:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        itemCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, itemCount, createdAt]
    BomItem:
      type: object
      properties:
        id:
          type: integer
        bomId:
          type: integer
        feederNumber:
          type: string
        partNumber:
          type: string
        description:
          type: string
        location:
          type: string
        quantity:
          type: integer
      required: [id, bomId, feederNumber, partNumber, quantity]
    BomDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/BomItem"
        createdAt:
          type: string
          format: date-time
      required: [id, name, items, createdAt]
    CreateBomRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required: [name]
    UpdateBomRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    CreateBomItemRequest:
      type: object
      properties:
        feederNumber:
          type: string
        partNumber:
          type: string
        description:
          type: string
        location:
          type: string
        quantity:
          type: integer
      required: [feederNumber, partNumber, quantity]
    UpdateBomItemRequest:
      type: object
      properties:
        feederNumber:
          type: string
        partNumber:
          type: string
        description:
          type: string
        location:
          type: string
        quantity:
          type: integer
...